Eliminate add-asset-html-webpack-plugin dependency
The add-asset-html-webpack-plugin was used by this project to insert link/script references in the project's generated index.html file. The links/scripts that were inserted pointed to the DLLs that this project generates. Since they're generated with a hash for long-term caching, we couldn't just hard-code the path names in the index file.

I was able to replace this behavior by using the templating system that the HtmlWebpackPlugin provides. This leads to a reduced amount of code in the webpack configuration file, as well as an index file that's a little more clearly documented.